feat(s3): posters and trailers have uploaded

+++ src/main/java/com/breeze/breezevideouser/DO/Reviews.java
@@ -30,7 +30,7 @@ public class Reviews implements Serializable {
     private Integer id;
 
     @ApiModelProperty("电影id")
-    private Long movieId;
+    private Integer movieId;
 
     @ApiModelProperty("评论作者")
     private String author;
